requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.32
langchain_openai==0.1.21
langchain_text_splitters==0.2.2
langgraph==0.2.3
python-dotenv==1.0.1
streamlit==1.37.1
faiss-cpu== 1.8.0.post1
pymupdf==1.24.9
numexpr==2.10.1
